requirements.txt
alabaster==0.7.12
arabic-reshaper==2.1.4
asgiref==3.4.1
asn1crypto==1.5.1
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.1.3
colorama==0.4.4
cryptography==36.0.0
cssselect2==0.7.0
cssutils==2.4.0
decorator==5.1.0
Django==3.2.8
django-codemirror2==0.2
django-cors-headers==3.11.0
django-cte==1.1.4
django-environ==0.8.1
django-extensions==3.1.3
django-filter==21.1
django-fsm==2.8.0
django-inlinecss==0.3.0
django-jazzmin==2.4.8
django-js-asset==1.2.2
django-mptt==0.9.1
django-simple-history==3.0.0
djangorestframework==3.12.4
docutils==0.17.1
drf-writable-nested==0.7.0
future==0.18.2
google-api-core==2.3.0
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
graphviz==0.18.2
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
imagesize==1.3.0
importlib-metadata==4.8.2
ipython==7.29.0
jedi==0.18.0
Jinja2==3.0.3
lxml==4.9.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
oauthlib==3.1.1
oscrypto==1.3.0
packaging==21.3
parso==0.8.2
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.22
protobuf==3.19.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydotplus==2.0.2
Pygments==2.10.0
pyHanko==0.15.0
pyhanko-certvalidator==0.19.5
pynliner==0.8.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
qrcode==7.3.1
reportlab==3.6.11
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.1
sqlparse==0.4.2
svglib==1.4.1
tinycss2==1.2.1
tqdm==4.64.1
traitlets==5.1.1
typing-extensions==3.10.0.2
tzdata==2022.5
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
xhtml2pdf==0.2.8
zipp==3.6.0
